LayoutTests/workers/sab/growable-shared-array-buffer-serialization-expected.txt

Growable SharedArrayBuffers should be serializable

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS arrayBuffer.growable is true
PASS arrayBuffer.byteLength is 36
PASS arrayBuffer.maxByteLength is 128
PASS result.growable is true
PASS result.byteLength is 36
PASS result.maxByteLength is 128
PASS result.growable is true
PASS result.byteLength is 128
PASS result.maxByteLength is 128
PASS arrayBuffer.growable is true
PASS arrayBuffer.byteLength is 128
PASS arrayBuffer.maxByteLength is 128
PASS successfullyParsed is true

TEST COMPLETE
